+++ src/cli.rs
@@ -13,9 +13,9 @@ pub fn make_app() -> Config {
     ];
     /// The pairings of impls to run. LHS calls RHS.
     static DEFAULT_TEST_PAIRS: &[(&str, &str)] = &[
-        (ABI_IMPL_RUSTC, ABI_IMPL_CC), // Rust calls C
-        (ABI_IMPL_CC, ABI_IMPL_RUSTC), // C calls Rust
-        (ABI_IMPL_CC, ABI_IMPL_CC),    // C calls C
+        (ABI_IMPL_RUSTC, ABI_IMPL_RUSTC), // (ABI_IMPL_RUSTC, ABI_IMPL_CC), // Rust calls C
+                                          // (ABI_IMPL_CC, ABI_IMPL_RUSTC), // C calls Rust
+                                          // (ABI_IMPL_CC, ABI_IMPL_CC),    // C calls C
     ];
 
     let app = clap::Command::new("abi-cafe")
